+++ examples/resources/dbtcloud_job/resource.tf
@@ -20,6 +20,6 @@ resource "dbtcloud_job" "test" {
     "schedule" : false
   }
   # this is the default that gets set up when modifying jobs in the UI
-  schedule_days = [0,1,2,3,4,5,6]
+  schedule_days = [0, 1, 2, 3, 4, 5, 6]
   schedule_type = "days_of_week"
 }