+++ examples/resources/dbtcloud_project_artefacts/resource.tf
@@ -1,0 +1,8 @@
+// use dbt_cloud_project_artefacts instead of dbtcloud_project_artefacts for the legacy resource names
+// legacy names will be removed from 0.3 onwards
+
+resource "dbtcloud_project_artefacts" "my_project_artefacts" {
+  project_id       = dbtcloud_project.my_project.id
+  docs_job_id      = dbtcloud_job.my_job.id
+  freshness_job_id = dbtcloud_job.my_job.id
+}