+++ examples/resources/dbtcloud_webhook/resource.tf
@@ -2,9 +2,9 @@
 // legacy names will be removed from 0.3 onwards
 
 resource "dbtcloud_webhook" "test_webhook" {
-  name = "my-webhook"
+  name        = "my-webhook"
   description = "My webhook"
-  client_url = "http://localhost/nothing"
+  client_url  = "http://localhost/nothing"
   event_types = [
     "job.run.started",
     "job.run.completed"