+++ pkg/resources/bigquery_credential_acceptance_test.go
@@ -104,7 +104,7 @@ func testAccCheckDbtCloudBigQueryCredentialDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("BigQuery credential still exists")
 		}
-		notFoundErr := "did not find"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)