+++ pkg/resources/group.go
@@ -3,6 +3,7 @@ package resources
 import (
 	"context"
 	"strconv"
+	"strings"
 
 	"github.com/dbt-labs/terraform-provider-dbtcloud/pkg/dbt_cloud"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
@@ -153,6 +154,10 @@ func resourceGroupRead(ctx context.Context, d *schema.ResourceData, m interface{
 
 	group, err := c.GetGroup(groupID)
 	if err != nil {
+		if strings.HasPrefix(err.Error(), "resource-not-found") {
+			d.SetId("")
+			return diags
+		}
 		return diag.FromErr(err)
 	}