+++ pkg/resources/group_acceptance_test.go
@@ -140,7 +140,7 @@ func testAccCheckDbtCloudGroupDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("Group still exists")
 		}
-		notFoundErr := "not found"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)