+++ pkg/resources/job_acceptance_test.go
@@ -251,7 +251,7 @@ func testAccCheckDbtCloudJobDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("Job still exists")
 		}
-		notFoundErr := "not found"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)