+++ pkg/resources/project_connection_acceptance_test.go
@@ -146,7 +146,7 @@ func testAccCheckDbtCloudProjectConnectionDestroy(s *terraform.State) error {
 		if project != nil {
 			return fmt.Errorf("Project still exists")
 		}
-		notFoundErr := "not found"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)