+++ pkg/resources/service_token_acceptance_test.go
@@ -142,7 +142,7 @@ func testAccCheckDbtCloudServiceTokenDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("ServiceToken still exists")
 		}
-		notFoundErr := "not found"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)