+++ pkg/resources/webhook_acceptance_test.go
@@ -151,7 +151,7 @@ func testAccCheckDbtCloudWebhookDestroy(s *terraform.State) error {
 		if err == nil {
 			return fmt.Errorf("Webhook still exists")
 		}
-		notFoundErr := "not found"
+		notFoundErr := "resource-not-found"
 		expectedErr := regexp.MustCompile(notFoundErr)
 		if !expectedErr.Match([]byte(err.Error())) {
 			return fmt.Errorf("expected %s, got %s", notFoundErr, err)